pyproject.toml
[project]
name = "omnis-installer"
version = "0.1.0"
description = "Next-generation, modular, and universal Linux installer"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
requires-python = ">=3.11"
authors = [
    { name = "GLF Team" }
]
keywords = ["linux", "installer", "qt6", "pyside6", "calamares-alternative"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Installation/Setup",
    "Typing :: Typed",
]

dependencies = [
    "PySide6>=6.6.0",
    "pydantic>=2.5.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-qt>=4.2.0",
    "mypy>=1.7.0",
    "ruff>=0.1.6",
    "types-PyYAML>=6.0",
]

[project.scripts]
omnis = "omnis.main:main"

[project.urls]
Homepage = "https://github.com/glmusic/Omnis"
Repository = "https://github.com/glmusic/Omnis.git"
Issues = "https://github.com/glmusic/Omnis/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/omnis"]

# =============================================================================
# Ruff - Linting & Formatting
# =============================================================================
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["omnis"]

# =============================================================================
# MyPy - Static Type Checking
# =============================================================================
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "PySide6.*",
]
ignore_missing_imports = true

# =============================================================================
# Pytest
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/omnis"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
